requirements.txt
agate==1.9.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
airflow-dbt==0.4.0
alembic==1.14.1
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-google==12.0.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.1
argcomplete==3.5.3
asgiref==3.8.1
astroid==3.3.8
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
black==25.1.0
blinker==1.9.0
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.5.2
cattrs==24.1.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.1
daff==1.3.46
db-dtypes==1.4.1
dbt-adapters==1.14.0
dbt-bigquery==1.9.1
dbt-common==1.15.0
dbt-core==1.9.2
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.7.4
decorator==5.2.1
deepdiff==7.0.1
Deprecated==1.2.18
dill==0.3.9
distlib==0.3.9
dnspython==2.7.0
docstring_parser==0.16
dotenv==0.9.9
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.17.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2025.2.0
gcloud-aio-auth==5.3.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.3.0
gcsfs==2025.2.0
google-ads==25.1.0
google-analytics-admin==0.23.4
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.81.0
google-cloud-alloydb==0.4.3
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.0
google-cloud-automl==2.16.1
google-cloud-batch==0.17.34
google-cloud-bigquery==3.29.0
google-cloud-bigquery-datatransfer==3.19.0
google-cloud-bigquery-storage==2.28.0
google-cloud-bigtable==2.28.1
google-cloud-build==3.31.0
google-cloud-compute==1.26.0
google-cloud-container==2.56.0
google-cloud-core==2.4.2
google-cloud-datacatalog==3.25.1
google-cloud-dataflow-client==0.8.16
google-cloud-dataform==0.5.16
google-cloud-dataplex==2.7.1
google-cloud-dataproc==5.18.0
google-cloud-dataproc-metastore==1.18.1
google-cloud-dlp==3.28.0
google-cloud-kms==3.4.0
google-cloud-language==2.17.0
google-cloud-logging==3.11.4
google-cloud-memcache==1.12.0
google-cloud-monitoring==2.27.0
google-cloud-orchestration-airflow==1.17.0
google-cloud-os-login==2.17.0
google-cloud-pubsub==2.28.0
google-cloud-redis==2.18.0
google-cloud-resource-manager==1.14.1
google-cloud-run==0.10.16
google-cloud-secret-manager==2.23.1
google-cloud-spanner==3.52.0
google-cloud-speech==2.31.0
google-cloud-storage==2.19.0
google-cloud-storage-transfer==1.16.0
google-cloud-tasks==2.19.1
google-cloud-texttospeech==2.25.0
google-cloud-translate==3.20.1
google-cloud-videointelligence==2.16.0
google-cloud-vision==3.10.0
google-cloud-workflows==1.17.0
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpc-interceptor==0.15.4
grpcio==1.70.0
grpcio-gcp==0.2.2
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
identify==2.6.8
idna==3.10
immutabledict==4.2.1
importlib-metadata==6.11.0
inflect==7.5.0
inflection==0.5.1
isodate==0.6.1
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
json-merge-patch==0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.10.0
leather==0.4.0
limits==4.0.1
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==25.2.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mashumaro==3.14
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
ordered-set==4.1.0
packaging==24.2
pandas==2.1.4
pandas-gbq==0.28.0
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.0.0
pip-tools==7.4.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
prison==0.2.1
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydata-google-auth==1.9.1
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.4
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.23.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
setproctitle==1.3.5
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
soda-core==3.5.0
soda-core-bigquery==3.5.0
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.1
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.8.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
tomli==2.2.1
tomlkit==0.13.2
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.2
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
